+++ clove/network/bitcoin_based/greencoin.py
@@ -1,9 +1,8 @@
 
-from clove.block_explorer.cryptoid import CryptoidAPI
-from clove.network.bitcoin.base import BitcoinBaseNetwork
+from clove.network.bitcoin.base import BitcoinBaseNetwork, NoAPI
 
 
-class Greencoin(CryptoidAPI, BitcoinBaseNetwork):
+class Greencoin(NoAPI, BitcoinBaseNetwork):
     """
     Class with all the necessary GRE network information based on
     http://www.github.com/greencoin-dev/GreenCoinV2/blob/master/src/chainparams.cpp