+++ clove/network/bitcoin_based/nevacoin.py
@@ -1,9 +1,7 @@
-
-from clove.block_explorer.cryptoid import CryptoidAPI
 from clove.network.bitcoin.base import BitcoinBaseNetwork, NoAPI
 
 
-class NevaCoin(CryptoidAPI, BitcoinBaseNetwork):
+class NevaCoin(NoAPI, BitcoinBaseNetwork):
     """
     Class with all the necessary NEVA network information based on
     http://www.github.com/Nevacoin/nevacoin/blob/master/src/chainparams.cpp
@@ -23,7 +21,7 @@ class NevaCoin(CryptoidAPI, BitcoinBaseNetwork):
     source_code_url = 'http://www.github.com/Nevacoin/nevacoin/blob/master/src/chainparams.cpp'
 
 
-class NevaCoinTestNet(NoAPI, NevaCoin):
+class NevaCoinTestNet(NevaCoin):
     """
     Class with all the necessary NEVA testing network information based on
     http://www.github.com/Nevacoin/nevacoin/blob/master/src/chainparams.cpp