+++ clove/network/bitcoin_based/peercoin.py
@@ -1,8 +1,7 @@
-from clove.block_explorer.cryptoid import CryptoidAPI
 from clove.network.bitcoin.base import BitcoinBaseNetwork, NoAPI
 
 
-class Peercoin(CryptoidAPI, BitcoinBaseNetwork):
+class Peercoin(NoAPI, BitcoinBaseNetwork):
     """
     Class with all the necessary Peercoin (PPC) network information based on
     https://github.com/peercoin/peercoin/blob/master/src/net.cpp
@@ -21,7 +20,7 @@ class Peercoin(CryptoidAPI, BitcoinBaseNetwork):
     source_code_url = 'https://github.com/peercoin/peercoin/blob/master/src/net.cpp'
 
 
-class PeercoinTestNet(NoAPI, Peercoin):
+class PeercoinTestNet(Peercoin):
     """
     Class with all the necessary Peercoin (PPC) testing network information based on
     https://github.com/peercoin/peercoin/blob/master/src/net.cpp