+++ clove/network/bitcoin_based/tao.py
@@ -1,8 +1,7 @@
-from clove.block_explorer.cryptoid import CryptoidAPI
 from clove.network.bitcoin.base import BitcoinBaseNetwork, NoAPI
 
 
-class Tao(CryptoidAPI, BitcoinBaseNetwork):
+class Tao(NoAPI, BitcoinBaseNetwork):
     """
     Class with all the necessary Tao (XTO) network information based on
     https://github.com/taoblockchain/tao-core/blob/master/src/chainparams.cpp
@@ -22,7 +21,7 @@ class Tao(CryptoidAPI, BitcoinBaseNetwork):
     source_code_url = 'https://github.com/taoblockchain/tao-core/blob/master/src/chainparams.cpp'
 
 
-class TaoTestNet(NoAPI, Tao):
+class TaoTestNet(Tao):
     """
     Class with all the necessary Tao (XTO) testing network information based on
     https://github.com/taoblockchain/tao-core/blob/master/src/chainparams.cpp